importing from master

+++ app/src/main/java/lu/fisch/canze/actors/Field.java
@@ -18,7 +18,7 @@ public class Field {
 
     private int from;
     private int to;
-    private int offset;
+    private double offset;
     private int id;
     private int divider;
     private int multiplier;
@@ -34,7 +34,7 @@ public class Field {
     private double value = 0;
     private int skipsCount = 0;
     
-    public Field(int id, int from, int to, int divider, int multiplier, int offset, int decimals, String format, String unit, String requestId, String responseId, String description, int car, int skips) {
+    public Field(int id, int from, int to, int divider, int multiplier, double offset, int decimals, String format, String unit, String requestId, String responseId, String description, int car, int skips) {
         this.from=from;
         this.to=to;
         this.offset=offset;
@@ -183,7 +183,7 @@ public class Field {
         this.to = to;
     }
 
-    public int getOffset() {
+    public double getOffset() {
         return offset;
     }
